Move tests to relevant files
